[REFACTOR] dto 패키지 리펙터링
