+++ src/main/java/umc/spring/converter/MemberConverter.java
@@ -1,10 +1,9 @@
 package umc.spring.converter;
 
-import jakarta.persistence.Converter;
 import umc.spring.domain.Member;
 import umc.spring.domain.enums.Gender;
-import umc.spring.web.dto.MemberRequestDTO;
-import umc.spring.web.dto.MemberResponseDTO;
+import umc.spring.web.dto.MemberDto.MemberRequestDTO;
+import umc.spring.web.dto.MemberDto.MemberResponseDTO;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
