+++ src/main/java/umc/spring/converter/ReviewConverter.java
@@ -3,14 +3,10 @@ package umc.spring.converter;
 import umc.spring.domain.Member;
 import umc.spring.domain.Review;
 import umc.spring.domain.Store;
-import umc.spring.domain.enums.Gender;
-import umc.spring.web.dto.MemberRequestDTO;
-import umc.spring.web.dto.MemberResponseDTO;
-import umc.spring.web.dto.ReviewRequestDTO;
-import umc.spring.web.dto.ReviewResponseDTO;
+import umc.spring.web.dto.ReviewDto.ReviewRequestDTO;
+import umc.spring.web.dto.ReviewDto.ReviewResponseDTO;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 
 public class ReviewConverter {
     public static ReviewResponseDTO.ReviewResultDTO toReviewResultDTO(Review review){
