+++ src/main/java/umc/spring/service/MemberService/MemberCommandService.java
@@ -1,7 +1,7 @@
 package umc.spring.service.MemberService;
 
 import umc.spring.domain.Member;
-import umc.spring.web.dto.MemberRequestDTO;
+import umc.spring.web.dto.MemberDto.MemberRequestDTO;
 
 public interface MemberCommandService {
     public Member joinMember(MemberRequestDTO.JoinDto request);
