+++ src/main/java/umc/spring/service/ReviewService/ReviewCommandService.java
@@ -2,7 +2,7 @@ package umc.spring.service.ReviewService;
 
 import org.springframework.transaction.annotation.Transactional;
 import umc.spring.domain.Review;
-import umc.spring.web.dto.ReviewRequestDTO;
+import umc.spring.web.dto.ReviewDto.ReviewRequestDTO;
 
 public interface ReviewCommandService {
     @Transactional
