+++ src/main/java/umc/spring/service/StoreService/StoreCommandService.java
@@ -2,7 +2,7 @@ package umc.spring.service.StoreService;
 
 import org.springframework.transaction.annotation.Transactional;
 import umc.spring.domain.Store;
-import umc.spring.web.dto.StoreRequestDTO;
+import umc.spring.web.dto.StoreDto.StoreRequestDTO;
 
 public interface StoreCommandService {
     @Transactional
