+++ src/main/java/umc/spring/service/StoreService/StoreCommandServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.spring.converter.StoreConverter;
 import umc.spring.domain.Store;
 import umc.spring.repository.StoreRepository;
-import umc.spring.web.dto.StoreRequestDTO;
+import umc.spring.web.dto.StoreDto.StoreRequestDTO;
 
 @Service
 @RequiredArgsConstructor
