+++ src/main/java/umc/spring/web/controller/ReviewRestController.java
@@ -10,8 +10,8 @@ import umc.spring.apiPayload.ApiResponse;
 import umc.spring.converter.ReviewConverter;
 import umc.spring.domain.Review;
 import umc.spring.service.ReviewService.ReviewCommandServiceImpl;
-import umc.spring.web.dto.ReviewRequestDTO;
-import umc.spring.web.dto.ReviewResponseDTO;
+import umc.spring.web.dto.ReviewDto.ReviewRequestDTO;
+import umc.spring.web.dto.ReviewDto.ReviewResponseDTO;
 
 @RestController
 @RequiredArgsConstructor
