+++ src/main/java/umc/spring/web/controller/TempRestController.java
@@ -1,7 +1,6 @@
 package umc.spring.web.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -9,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import umc.spring.apiPayload.ApiResponse;
 import umc.spring.converter.TempConverter;
 import umc.spring.service.TempService.TempQueryService;
-import umc.spring.web.dto.TempResponse;
+import umc.spring.web.dto.TempDto.TempResponse;
 
 @RestController
 @RequestMapping("/temp")
