+++ src/main/java/umc/spring/web/dto/MemberDto/MemberRequestDTO.java
@@ -1,4 +1,4 @@
-package umc.spring.web.dto;
+package umc.spring.web.dto.MemberDto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
