+++ src/main/java/umc/spring/web/dto/StoreDto/StoreRequestDTO.java
@@ -1,4 +1,4 @@
-package umc.spring.web.dto;
+package umc.spring.web.dto.StoreDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
