+++ src/main/java/umc/spring/web/dto/TempDto/TempRequest.java
@@ -1,0 +1,4 @@
+package umc.spring.web.dto.TempDto;
+
+public class TempRequest {
+}
