+++ src/main/java/umc/spring/web/dto/TempDto/TempResponse.java
@@ -1,4 +1,4 @@
-package umc.spring.web.dto;
+package umc.spring.web.dto.TempDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
