--- src/main/java/umc/spring/web/dto/TempRequest.java
@@ -1,4 +1,0 @@
-package umc.spring.web.dto;
-
-public class TempRequest {
-}
